requirements.txt
django-ckeditor
Pillow